S41_Ep02_Tweets
This code was used to collect and analyze tweets contained in the paper "Don’t Make My Entertainment Political! Social Media Responses to Narratives of Racial Duty on Competitive Reality Television Series." These tweets were published in response to an episode announcement tweet made via CBS' Twitter account the week each of the following episodes from Survivor 41 originally aired: Episode 1, Episode 5, Episode 6, and Episode 11.  Given the sensitive nature of tweets and the inability to fully de-identify tweets, the complete tweet data are not posted here. However, for those interested in collecting the data themselves and reproducing the analysis, the code used to collect the data is included. In the analysis, tweet data were cleaned to remove tweets that only contained emojis, hashtags, or @ replies. Tweets published in languages other than English were also removed. The inclusion and exclusion criteria for tweets are detailed in the published paper and appendices.