Add files via upload
This R script file contains the code used to collect original tweet data and conduct the sentiment analysis contained in the paper "Don’t Make My Entertainment Political! Social Media Responses to Narratives of Racial Duty on Competitive Reality Television Series." These tweets were published in response to an episode announcement tweet made via CBS' Twitter account the week each episode aired for the following episodes: 2, 5, 6, and 11. Given the sensitive nature of the content and my inability to fully de-identify users’ tweets, the complete dataset that was used in the analysis is not posted here. However, for researchers who wish to replicate the analysis themselves, the code that was used to pull tweets is included in the script file. The cleaned dataset removed tweets that only contained emojis or @replies. Tweets published in languages other than English were also removed. The inclusion/exclusion criteria used in the study is described in the article main text and appendix.